pyproject.toml
[build-system]
requires = ["setuptools>=68","wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whirlwind"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
  "rich>=13",
  "numpy",
  "rasterio",
  "pyyaml",
  "cmd2>=2.4",
]

[project.scripts]
W = "whirlwind.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where=["src"]
